move activations
