+++ docs/source/api_ref_training.rst
@@ -14,3 +14,17 @@ Utilities for working in a reduced precision setting.
     :nosignatures:
 
     get_quantizer_mode
+
+
+.. _ac_label:
+
+Memory Management
+-----------------
+
+Utilities to reduce memory consumption during training.
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    apply_selective_activation_checkpointing
