+++ recipes/full_finetune_distributed.py
@@ -24,8 +24,8 @@ from torchtune import config, modules, utils
 from torchtune.data import padded_collate
 from torchtune.datasets import ConcatDataset
 from torchtune.recipe_interfaces import FTRecipeInterface
+from torchtune.training.activations import apply_selective_activation_checkpointing
 from torchtune.utils import DummyProfiler, PROFILER_KEY
-from torchtune.utils.activations import apply_selective_activation_checkpointing
 
 from tqdm import tqdm
