+++ torchtune/training/__init__.py
@@ -4,8 +4,10 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+from torchtune.training.activations import apply_selective_activation_checkpointing
 from torchtune.training.quantization import get_quantizer_mode
 
 __all__ = [
+    "apply_selective_activation_checkpointing",
     "get_quantizer_mode",
 ]
